+++ .github/workflows/tests.yml
@@ -70,6 +70,7 @@ jobs:
             cuda_version: 11.8.0
             python_version: "3.10"
             pytorch: 2.1.2
+            axolotl_args: "--extra-index-url https://download.pytorch.org/whl/cu118"
           - cuda: 121
             cuda_version: 12.1.0
             python_version: "3.10"
@@ -87,11 +88,13 @@ jobs:
           # Set up build arguments
           BASE_TAG="main-base-py${{ matrix.python_version }}-cu${{ matrix.cuda }}-${{ matrix.pytorch }}"
           CUDA="${{ matrix.cuda }}"
+          AXOLOTL_ARGS="${{ matrix.axolotl_args }}"
           PYTORCH_VERSION="${{ matrix.pytorch }}"
           # Build the Docker image
           docker build . \
             --file ./docker/Dockerfile-tests \
             --build-arg BASE_TAG=$BASE_TAG \
+            --build-arg AXOLOTL_ARGS="$AXOLOTL_ARGS" \
             --build-arg CUDA=$CUDA \
             --build-arg GITHUB_REF=$GITHUB_REF \
             --build-arg PYTORCH_VERSION=$PYTORCH_VERSION \
